ai-service/requirements.txt
fastapi
uvicorn
python-multipart
spacy
scikit-learn
PyMuPDF
python-docx
pandas
numpy